requirements.txt
asgiref==3.3.4
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
coverage==5.5
deprecation==2.1.0
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-filter==21.1
djangorestframework==3.12.4
eyed3==0.9.6
filelock==3.1.0
filetype==1.0.8
future==0.18.2
gunicorn==20.1.0
idna==3.3
packaging==21.0
Pillow==8.3.2
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.7.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
serializers==0.2.4
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
virtualenv==20.8.1
virtualenvwrapper-win==1.2.6
whitenoise==5.3.0
